pdf-converter/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
anthropic==0.40.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
psycopg2-binary==2.9.9
google-generativeai
openai
pypdf
